refactor: rename types; extract into separate file
