+++ src/lib/openapi/index.ts
@@ -104,7 +104,7 @@ import { groupSchema } from './spec/group-schema';
 import { groupsSchema } from './spec/groups-schema';
 import { groupUserModelSchema } from './spec/group-user-model-schema';
 import { usersGroupsBaseSchema } from './spec/users-groups-base-schema';
-import { OpenApiTags } from './util/api-operation';
+import { openApiTags } from './util/openapi-tags';
 import { searchEventsSchema } from './spec/search-events-schema';
 
 // All schemas in `openapi/spec` should be listed here.
@@ -272,6 +272,6 @@ export const createOpenApiSchema = ({
             },
             schemas: mapValues(schemas, removeJsonSchemaProps),
         },
-        tags: OpenApiTags,
+        tags: openApiTags,
     };
 };
